+++ src/pages/home/home.config.js
