+++ src/pages/search/search.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import { View } from '@tarojs/components'
+import TabBar from '../../component/tabBar/tabBar'
+
+const Search = () => {
+  return (
+    <View>
+      <View>搜索</View>
+      <TabBar value={1} />
+    </View>
+  )
+}
+
+export default Search
